% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21012.R
\name{IPTWmethod}
\alias{IPTWmethod}
\title{Adjusting selection bias and misclassification by IPTW.}
\usage{
IPTWmethod(dt, Treat, Y, ehat, s01, s11, p01, p11, strata = F)
}
\arguments{
\item{dt}{The dataset contains treatment, outcome, covariates, and propensity score.}

\item{Treat}{Treatment}

\item{Y}{outcome}

\item{ehat}{propensity score}

\item{s01}{selection parameter P(S=1|Y=0)}

\item{s11}{selection parameter P(S=1|Y=1)}

\item{p01}{misclassification parameter P(Y*=1|Y=0)}

\item{p11}{misclassification parameter P(Y*=1|Y=1)}

\item{strata}{strata data according to covariates}
}
\value{
adjusted ATE
}
\description{
Adjusting selection bias and misclassification by IPTW.
}
\examples{
\dontrun{
data(dt)
IPTWmethod(dt, Treat, Y, ehat, 1, 1, 0, 1)
}
}
