% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21012.R
\name{paramethod}
\alias{paramethod}
\title{Adjusting selection bias and misclassification by Spline method.}
\usage{
paramethod(object, s01, s11, p01, p11, Boot = 1000)
}
\arguments{
\item{object}{The outcome model}

\item{s01}{selection parameter P(S=1|Y=0)}

\item{s11}{selection parameter P(S=1|Y=1)}

\item{p01}{misclassification parameter P(Y*=1|Y=0)}

\item{p11}{misclassification parameter P(Y*=1|Y=1)}

\item{Boot}{MCmethod for eastimating the integration, size of the MC sample.}
}
\value{
adjusted ATE
}
\description{
Adjusting selection bias and misclassification by Spline method.
}
\examples{
\dontrun{
data(dt)
gfit <- glm(formula = Y~X1+X2+X3+Treat, family = binomial(link = "logit"), data = dt)
paramethod(gfit, 1, 1, 0, 1)
}
}
