% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21012.R
\name{splinemethod}
\alias{splinemethod}
\title{Adjusting selection bias and misclassification by Spline method.}
\usage{
splinemethod(dt, s01, s11, p01, p11, df = 10, Boot = 1000)
}
\arguments{
\item{dt}{The dataset contains treatment, outcome, covariates, and propensity score.}

\item{s01}{selection parameter P(S=1|Y=0)}

\item{s11}{selection parameter P(S=1|Y=1)}

\item{p01}{misclassification parameter P(Y*=1|Y=0)}

\item{p11}{misclassification parameter P(Y*=1|Y=1)}

\item{df}{the degree of B-spline}

\item{Boot}{MCmethod for eastimating the integration, size of the MC sample.}
}
\value{
adjusted ATE
}
\description{
Adjusting selection bias and misclassification by Spline method.
}
\examples{
\dontrun{
data(dt)
splinemethod(dt, 1, 1, 0, 1)
}
}
