\name{ATEms-package}
\alias{ATEms-package}
\docType{package}
\title{
  Estimating Average Treatment Effect with misclassification and selection bias.
}
\description{
  Reducing the bias caused by misclassification and selection bias in estimating Average Treatment Effect(ATE). Four methods are offered for chosen.
}
\author{
Min Zeng

Maintainer: Min Zeng <coolzg@mail.ustc.edu.cn>
}
\keyword{ package }
